idx type Long으로 변경 agency #41

+++ src/main/java/com/tsp/new_tsp_front/api/agency/domain/FrontAgencyEntity.java
@@ -32,7 +32,7 @@ public class FrontAgencyEntity extends NewCommonMappedClass {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "idx")
-    private Integer idx;
+    private Long idx;
 
     @Column(name = "agency_name")
     @NotEmpty(message = "소속사명 입력은 필수입니다.")
